Add APi payment Bug

+++ 8.4.0/aspnet-core/src/ECMS.Application/Payment/Dto/VnPaymentRequestDto.cs
@@ -8,10 +8,10 @@ namespace ECMS.Payment.Dto
 {
     public class VnPaymentRequestDto
     {
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
         public string FullName { get; set; }
         public string Description { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public DateTime CreatedDate { get; set; }
     }
 }
